requirements.txt
aerich==0.9.1
aiofiles==24.1.0
aiogram==3.21.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
arq==0.26.3
asyncclick==8.1.8.0
asyncpg==0.30.0
attrs==25.3.0
certifi==2025.7.14
click==8.2.1
colorama==0.4.6
dictdiffer==0.9.0
frozenlist==1.7.0
hiredis==3.2.1
idna==3.10
iso8601==2.1.0
magic-filter==1.0.12
multidict==6.6.3
propcache==0.3.2
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic-settings-yaml==0.2.0
pydantic_core==2.33.2
PyJWT==2.10.1
pypika-tortoise==0.6.1
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
redis==5.3.1
sniffio==1.3.1
tomlkit==0.13.3
tortoise-orm==0.25.1
typing-inspection==0.4.1
typing_extensions==4.14.1
yarl==1.20.1
